UwU refactor
Stores valid chars as a 'word' which gets processed at its end, and
additionally added a new 'smarter' method to repwace that matches any
cAsE vArIaNcE if the substitution is the same length as the pattern.
Complemented with a new titlecase fn, this should completely fix all
scenarios of words that should be replaced not actually being replaced,
with the exception of replacing varied len replacements on words with
gOOfY case, as there's no "correct" solution to that.